+++ src/etools/applications/tokens/templates/tokens/login.html
@@ -83,7 +83,7 @@
     </div>
     <div class="card-content">
       <h3>Sign in with Email</h3>
-      <form id="form" method="POST" action="{% url "email_auth:login" %}">
+      <form id="form" method="POST" action="{% url 'tokens:login' %}">
         {% csrf_token %}
         {% for error in form.non_field_errors %}
           {% if error %}
