+++ src/etools/applications/tokens/utils.py
@@ -1,4 +1,3 @@
-
 from django.conf import settings
 from django.urls import reverse
 
@@ -26,4 +25,7 @@ def update_url_with_auth_token(url, user):
 
 
 def get_token_auth_link(user):
-    return update_url_with_auth_token(urljoin(site_url(), reverse('email_auth:login')), user)
+    return update_url_with_auth_token(
+        urljoin(site_url(), reverse('tokens:login')),
+        user
+    )
