+++ src/etools/config/settings/base.py
@@ -115,7 +115,7 @@ MIDDLEWARE = (
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.contrib.auth.middleware.SessionAuthenticationMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'etools.applications.email_auth.middleware.TokenAuthenticationMiddleware',
+    'etools.applications.tokens.middleware.TokenAuthenticationMiddleware',
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
@@ -195,7 +195,7 @@ SHARED_APPS = (
     'etools.applications.utils.writable_serializers',
     'etools.applications.utils.permissions',
     'waffle',
-    'etools.applications.email_auth',
+    'etools.applications.tokens',
     'etools.applications.permissions2',
 )
 TENANT_APPS = (
