+++ src/etools/config/urls.py
@@ -54,7 +54,7 @@ urlpatterns = [
     # Used for admin and dashboard pages in django
     url(r'^$', ModuleRedirectView.as_view(), name='dashboard'),
     url(r'^login/$', MainView.as_view(), name='main'),
-    url(r'^email-auth/', include('etools.applications.email_auth.urls')),
+    url(r'^tokens/', include('etools.applications.tokens.urls')),
 
     url(r'^api/static_data/$', StaticDataView.as_view({'get': 'list'}), name='public_static'),
